Move accounts.helpers code to accounts.utils
