+++ src/olympia/accounts/tests/test_utils.py
@@ -8,7 +8,7 @@ from django.test.utils import override_settings
 
 import mock
 
-from olympia.accounts import helpers
+from olympia.accounts import utils
 
 FXA_CONFIG = {
     'default': {
@@ -26,7 +26,7 @@ def test_fxa_config_anonymous():
     context = mock.MagicMock()
     context['request'].session = {'fxa_state': 'thestate!'}
     context['request'].user.is_authenticated.return_value = False
-    assert helpers.fxa_config(context) == {
+    assert utils.fxa_config(context) == {
         'clientId': 'foo',
         'state': 'thestate!',
         'oauthHost': 'https://accounts.firefox.com/oauth',
@@ -41,7 +41,7 @@ def test_fxa_config_logged_in():
     context['request'].session = {'fxa_state': 'thestate!'}
     context['request'].user.is_authenticated.return_value = True
     context['request'].user.email = 'me@mozilla.org'
-    assert helpers.fxa_config(context) == {
+    assert utils.fxa_config(context) == {
         'clientId': 'foo',
         'state': 'thestate!',
         'email': 'me@mozilla.org',
@@ -56,7 +56,7 @@ def test_default_fxa_login_url_with_state():
     path = '/en-US/addons/abp/?source=ddg'
     request = RequestFactory().get(path)
     request.session = {'fxa_state': 'myfxastate'}
-    raw_url = helpers.default_fxa_login_url(request)
+    raw_url = utils.default_fxa_login_url(request)
     url = urlparse.urlparse(raw_url)
     base = '{scheme}://{netloc}{path}'.format(
         scheme=url.scheme, netloc=url.netloc, path=url.path)
@@ -72,61 +72,61 @@ def test_default_fxa_login_url_with_state():
 
 
 @mock.patch(
-    'olympia.accounts.helpers.default_fxa_login_url',
+    'olympia.accounts.utils.default_fxa_login_url',
     lambda c: 'http://auth.ca')
-@mock.patch('olympia.accounts.helpers.waffle.switch_is_active')
+@mock.patch('olympia.accounts.utils.waffle.switch_is_active')
 def test_login_link_migrated(switch_is_active):
     switch_is_active.return_value = True
-    assert helpers.login_link({'request': mock.MagicMock()}) == (
+    assert utils.login_link({'request': mock.MagicMock()}) == (
         'http://auth.ca')
     switch_is_active.assert_called_with('fxa-migrated')
 
 
-@mock.patch('olympia.accounts.helpers.waffle.switch_is_active')
+@mock.patch('olympia.accounts.utils.waffle.switch_is_active')
 def test_login_link_not_migrated(switch_is_active):
     request = RequestFactory().get('/en-US/foo')
     switch_is_active.return_value = False
-    assert helpers.login_link({'request': request}) == (
+    assert utils.login_link({'request': request}) == (
         '{}?to=%2Fen-US%2Ffoo'.format(reverse('users.login')))
     switch_is_active.assert_called_with('fxa-migrated')
 
 
 @mock.patch(
-    'olympia.accounts.helpers.default_fxa_login_url',
+    'olympia.accounts.utils.default_fxa_login_url',
     lambda c: 'http://auth.ca')
-@mock.patch('olympia.accounts.helpers.waffle.switch_is_active')
+@mock.patch('olympia.accounts.utils.waffle.switch_is_active')
 def test_register_link_migrated(switch_is_active):
     switch_is_active.return_value = True
-    assert helpers.register_link({'request': mock.MagicMock()}) == (
+    assert utils.register_link({'request': mock.MagicMock()}) == (
         'http://auth.ca')
     switch_is_active.assert_called_with('fxa-migrated')
 
 
-@mock.patch('olympia.accounts.helpers.waffle.switch_is_active')
+@mock.patch('olympia.accounts.utils.waffle.switch_is_active')
 def test_register_link_not_migrated(switch_is_active):
     request = RequestFactory().get('/en-US/foo')
     switch_is_active.return_value = False
-    assert helpers.register_link({'request': request}) == (
+    assert utils.register_link({'request': request}) == (
         '{}?to=%2Fen-US%2Ffoo'.format(reverse('users.register')))
     switch_is_active.assert_called_with('fxa-migrated')
 
 
-@mock.patch('olympia.accounts.helpers.waffle.switch_is_active', lambda s: True)
+@mock.patch('olympia.accounts.utils.waffle.switch_is_active', lambda s: True)
 def test_unicode_next_path():
     path = u'/en-US/føø/bãr'
     request = RequestFactory().get(path)
     request.session = {}
-    url = helpers.login_link({'request': request})
+    url = utils.login_link({'request': request})
     state = urlparse.parse_qs(urlparse.urlparse(url).query)['state'][0]
     next_path = urlsafe_b64decode(state.split(':')[1] + '===')
     assert next_path.decode('utf-8') == path
 
 
-@mock.patch('olympia.accounts.helpers.default_fxa_login_url')
+@mock.patch('olympia.accounts.utils.default_fxa_login_url')
 def test_redirect_for_login(default_fxa_login_url):
     fxa_url = 'https://accounts.firefox.com/login'
     default_fxa_login_url.return_value = fxa_url
     request = mock.MagicMock()
-    response = helpers.redirect_for_login(request)
+    response = utils.redirect_for_login(request)
     default_fxa_login_url.assert_called_with(request)
     assert response['location'] == fxa_url
