+++ src/olympia/amo/decorators.py
@@ -14,7 +14,7 @@ from olympia.users.utils import get_task_user
 
 from . import models as context
 from .utils import JSONEncoder
-from olympia.accounts.helpers import redirect_for_login
+from olympia.accounts.utils import redirect_for_login
 
 
 task_log = commonware.log.getLogger('z.task')
